search: move job logic out of gqlbackend
