+++ cmd/frontend/graphqlbackend/search_results_stats_languages.go
@@ -12,6 +12,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
 	"github.com/sourcegraph/sourcegraph/internal/inventory"
+	"github.com/sourcegraph/sourcegraph/internal/search/job"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming"
 	"github.com/sourcegraph/sourcegraph/internal/types"
@@ -40,13 +41,13 @@ func (srs *searchResultsStats) Languages(ctx context.Context) ([]*languageStatis
 func (srs *searchResultsStats) getResults(ctx context.Context) (result.Matches, error) {
 	args := srs.sr.JobArgs()
 	srs.once.Do(func() {
-		job, err := toSearchJob(args, srs.sr.Query)
+		j, err := job.ToSearchJob(args, srs.sr.Query)
 		if err != nil {
 			srs.err = err
 			return
 		}
 		agg := streaming.NewAggregatingStream()
-		_, err = job.Run(ctx, srs.sr.db, agg)
+		_, err = j.Run(ctx, srs.sr.db, agg)
 		if err != nil {
 			srs.err = err
 			return
